rag-answer/requirements.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0